PMM-10246 moved fields to seperate components
