+++ public/app/percona/integrated-alerting/components/Table/Filter/components/fields/SearchTextField.tsx
@@ -1,0 +1,21 @@
+import { Input } from '@grafana/ui';
+import React from 'react';
+import { Field } from 'react-final-form';
+import { SEARCH_INPUT_FIELD_NAME } from '../../Filter.constants';
+import { Messages } from '../../Filter.messages';
+
+export const SearchTextField = () => {
+  return (
+    <Field name={SEARCH_INPUT_FIELD_NAME}>
+      {({ input }) => (
+        <Input
+          type="text"
+          placeholder={Messages.searchPlaceholder}
+          {...input}
+          data-testid={SEARCH_INPUT_FIELD_NAME}
+          autoFocus={true}
+        />
+      )}
+    </Field>
+  );
+};
